added outb and inb to common.h from vga and keyboard
