+++ kernel/common.c
@@ -1,6 +1,18 @@
 //common.c for all common tasks 
 //Jacob Mealey
 //
+#include "common.h"
+
+uint8_t inb(uint16_t port)
+{
+	uint8_t ret;
+	asm volatile("inb %1, %0" : "=a" (ret) : "dN" (port));
+	return ret;
+}
+void outb(uint16_t port, uint8_t value)
+{
+	    asm volatile ("outb %1, %0" : : "dN" (port), "a" (value));
+}
 
 int string_length(char* string)
 {
