+++ kernel/common.h
@@ -6,7 +6,12 @@
 #endif
 
 #include <stddef.h>
+#include <stdbool.h> /* C doesn't have booleans by default. */
+
+#include <stddef.h>
 #include <stdint.h>
+uint8_t inb(uint16_t port);
+void outb(uint16_t port, uint8_t value);
 int string_length(char* string);
 void concatc(char* a, char b, char* c);
 int string_compare(char* str1, char* str2);
