+++ kernel/keyboard.h
@@ -9,7 +9,6 @@
 #include <stdint.h>
 
 unsigned char kbdus[256];
-uint8_t inb(uint16_t);
 char getScancode();
 void handleKeys();
