+++ kernel/vga.h
@@ -6,7 +6,6 @@
 #endif
 #include <stddef.h>
 #include <stdint.h>
-void outb(uint16_t port, uint8_t value);
 void clear_screen(uint8_t color);
 
 void putchar(char c, uint8_t color);
